# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Cycling Fantasy Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stages/current:
    get:
      operationId: getCurrentStage
      tags: [stages]
      summary: Get the current or next upcoming stage
      responses:
        "200":
          description: Current stage info with transfer window status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentStageResponse"
        "404":
          description: No active stage found
  /stages/{id}:
    get:
      operationId: getStage
      tags: [stages]
      summary: Get stage details including results if completed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stage detail with results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StageDetail"
        "404":
          description: Stage not found

  # Team
  /team:
    get:
      operationId: getMyTeam
      tags: [team]
      summary: Get my current fantasy team
      responses:
        "200":
          description: Current team selection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyTeam"
        "401":
          description: Not authenticated
    put:
      operationId: updateMyTeam
      tags: [team]
      summary: Update my team selection (transfer window must be open)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamUpdate"
      responses:
        "200":
          description: Updated team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyTeam"
        "400":
          description: Invalid selection or transfer window closed
        "401":
          description: Not authenticated

  # Leaderboard
...
components:
...
  schemas:
...
    Stage:
      type: object
      required: [id, stageNumber, name, startCity, endCity, date, stageType, status, transferDeadline]
      properties:
        id:
          type: integer
        stageNumber:
          type: integer
        name:
          type: string
        startCity:
          type: string
        endCity:
          type: string
        date:
          type: string
          format: date
        stageType:
          type: string
          enum: [flat, hilly, mountain, time_trial, rest]
        status:
          type: string
          enum: [upcoming, transfer_closed, live, completed]
        transferDeadline:
          type: ["string", "null"]
          format: date-time
        pcsUrl:
          type: ["string", "null"]
    CurrentStageResponse:
      type: object
      required: [stage, transferWindowOpen, minutesUntilClose]
      properties:
        stage:
          $ref: "#/components/schemas/Stage"
        transferWindowOpen:
          type: boolean
        minutesUntilClose:
          type: ["integer", "null"]
    StageResult:
      type: object
      required: [id, stageId, riderId, riderName, proTeam, position, dnf, fantasyPoints]
      properties:
        id:
          type: integer
        stageId:
          type: integer
        riderId:
          type: integer
        riderName:
          type: string
        proTeam:
          type: string
        position:
          type: ["integer", "null"]
        dnf:
          type: boolean
        fantasyPoints:
          type: number
        pointsBreakdown:
          type: object
          properties:
            stage:
              type: number
            jerseys:
              type: number
            kom:
              type: number
            sprint:
              type: number
            combative:
              type: number
            penalty:
              type: number
          required: [stage, jerseys, kom, sprint, combative, penalty]
    StageDetail:
      type: object
      required: [stage, results]
      properties:
        stage:
          $ref: "#/components/schemas/Stage"
        results:
          type: array
          items:
            $ref: "#/components/schemas/StageResult"
    RiderSummary:
      type: object
      required: [id, name, proTeam, oddsDecimal, oddsLabel, currentJerseys, dnf]
      properties:
        id:
          type: integer
        name:
          type: string
        proTeam:
          type: string
        oddsDecimal:
          type: number
        oddsLabel:
          type: string
        photoUrl:
          type: ["string", "null"]
        currentJerseys:
          type: array
          items:
            type: string
        dnf:
          type: boolean
    MyTeam:
      type: object
      required: [riders, captainRiderId, transferWindowOpen]
      properties:
        riders:
          type: array
          items:
            $ref: "#/components/schemas/RiderSummary"
        captainRiderId:
          type: ["integer", "null"]
        transferWindowOpen:
          type: boolean
        totalPoints:
          type: number
    TeamUpdate:
      type: object
      required: [riderIds, captainRiderId]
      properties:
        riderIds:
          type: array
          items:
            type: integer
          minItems: 1
          maxItems: 8
        captainRiderId:
          type: integer
...