# lib/api-zod/src/generated/types/myTeam.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Cycling Fantasy Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { RiderSummary } from './riderSummary';

export interface MyTeam {
  riders: RiderSummary[];
  /** @nullable */
  captainRiderId: number | null;
  transferWindowOpen: boolean;
  totalPoints?: number;
}