# lib/api-zod/src/generated/types/stageResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Cycling Fantasy Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { StageResultPointsBreakdown } from './stageResultPointsBreakdown';

export interface StageResult {
  id: number;
  stageId: number;
  riderId: number;
  riderName: string;
  proTeam: string;
  /** @nullable */
  position: number | null;
  dnf: boolean;
  fantasyPoints: number;
  pointsBreakdown?: StageResultPointsBreakdown;
}