# lib/api-zod/src/generated/types/teamUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Cycling Fantasy Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface TeamUpdate {
  /**
     * @minItems 1
     * @maxItems 8
     */
  riderIds: number[];
  captainRiderId: number;
}